requirements.txt
streamlit
torch
torchvision
Pillow
numpy
pandas
h5py
timm
